Logger

+++ ApocalypseShared/Common/World/WorldConstants.h
@@ -9,4 +9,4 @@ constexpr int InventorySize = 9;
 
 constexpr int ChunkSize = 32;
 constexpr int ChunkVolume = ChunkSize * ChunkSize * ChunkSize;
-constexpr int BIG_NUMBER = 0x10000000000000;
+constexpr __int64 BIG_NUMBER = 0x10000000000000;
